switched rating data type from float to integer

+++ data/migrations/20191019030531_rvnb_schema.js
@@ -60,7 +60,7 @@ exports.up = function(knex) {
       .notNullable();
 
       property
-      .float('rating', 2, 2);
+      .integer('rating', 2, 2);
 
       property
       .integer('owner_id')
